chore: pull merkle tree implementation into `acvm-backend-barretenberg`
