+++ acvm_backend_barretenberg/src/lib.rs
@@ -12,6 +12,8 @@ mod acvm_interop;
 mod composer;
 #[cfg(all(feature = "native", test))]
 mod crs;
+#[cfg(test)]
+mod merkle;
 mod pedersen;
 mod pippenger;
 mod scalar_mul;
