+++ common/Cargo.toml
@@ -12,8 +12,6 @@ publish = false
 [dependencies]
 acvm = { version = "0.9.0", features = ["bn254"] }
 
-sled = "0.34.6"
-blake2 = "0.9.1"
 dirs = { version = "3.0", optional = true }
 reqwest = { version = "0.11.16", optional = true, default-features = false, features = ["stream", "rustls-tls"] }
 tokio = { version = "1.0", optional = true }
