+++ common/src/lib.rs
@@ -5,7 +5,6 @@ pub mod barretenberg_structures;
 
 #[cfg(feature = "std")]
 pub mod crs;
-pub mod merkle;
 
 // Re-export acvm
 pub use acvm;
